libc:machine:risc-v:unifying elf relocation code.
