+++ libs/libc/machine/risc-v/rv64/Make.defs
@@ -18,11 +18,3 @@
 #
 ############################################################################
 
-ifeq ($(CONFIG_LIBC_ARCH_ELF),y)
-
-CSRCS += arch_elf.c
-
-DEPPATH += --dep-path machine/risc-v/rv64
-VPATH += :machine/risc-v/rv64
-
-endif
